+++ backend/src/main/java/com/carffeine/carffeine/station/service/review/ReviewQueryService.java
@@ -1,7 +1,7 @@
 package com.carffeine.carffeine.station.service.review;
 
-import com.carffeine.carffeine.station.controller.review.dto.TotalRatingsResponse;
 import com.carffeine.carffeine.station.infrastructure.repository.review.ReviewQueryRepository;
+import com.carffeine.carffeine.station.infrastructure.repository.review.TotalRatingsResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
