+++ backend/src/test/java/com/carffeine/carffeine/station/integration/review/ReviewIntegrationTest.java
@@ -7,11 +7,11 @@ import com.carffeine.carffeine.member.domain.MemberRepository;
 import com.carffeine.carffeine.member.exception.MemberException;
 import com.carffeine.carffeine.member.exception.MemberExceptionType;
 import com.carffeine.carffeine.station.controller.review.dto.ReviewResponses;
-import com.carffeine.carffeine.station.controller.review.dto.TotalRatingsResponse;
 import com.carffeine.carffeine.station.domain.review.Review;
 import com.carffeine.carffeine.station.domain.review.ReviewRepository;
 import com.carffeine.carffeine.station.domain.station.Station;
 import com.carffeine.carffeine.station.domain.station.StationRepository;
+import com.carffeine.carffeine.station.infrastructure.repository.review.TotalRatingsResponse;
 import com.carffeine.carffeine.station.service.review.dto.CreateReviewRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
